[WIP] Hollestered all methods.
